i3-status: dont display weather widget if it failed
Sometimes the API used to fetch weather information likes to display
very long messages explaining things like the site going down for
maintaince. The bar would show as much of the message as it could,
bridging over my entire monitor.

Since the widget is just meant to display the temperature and an icon it
will now not do anything if the request returns more than 5 characters